+++ AngularClient/src/main/core/project-editor/project-editor.component.ts
@@ -50,6 +50,9 @@ export class ProjectEditorComponent implements OnDestroy, OnInit {
   subscriptions: Subscription[] = [];
   username = "";
 
+  // When items income decreased then push to array for each rounds
+  decreasedItems: ElementItem[] = [];
+
   // User Project Array
   projectDataSet = Array<Project>();
 
@@ -71,6 +74,7 @@ export class ProjectEditorComponent implements OnDestroy, OnInit {
   timerDelay = 1000;
   timerSubscription = observableTimer(1500, this.timerDelay).subscribe(() => {
     this.refreshPage();
+    this.changeStatus();
   });
 
   // QRCode
@@ -146,11 +150,12 @@ export class ProjectEditorComponent implements OnDestroy, OnInit {
   resetTimer(): void {
 
     this.project.resetRounds();
-
+    this.decreasedItems = []; //clear
     this.timerDelay = 1000;
     this.timerSubscription.unsubscribe();
     this.timerSubscription = observableTimer(1000, this.timerDelay).subscribe(() => {
       this.refreshPage();
+      this.changeStatus();
     });
 
     console.log(`Reset: Timer delay time set to 1 second..`);
@@ -163,6 +168,7 @@ export class ProjectEditorComponent implements OnDestroy, OnInit {
     this.timerSubscription.unsubscribe();
     this.timerSubscription = observableTimer(1000, this.timerDelay).subscribe(() => {
       this.refreshPage();
+      this.changeStatus();
     });
     console.log(`Timer delay time set to ${this.timerDelay / 1000} s.`);
   }
@@ -174,6 +180,7 @@ export class ProjectEditorComponent implements OnDestroy, OnInit {
     this.timerSubscription.unsubscribe();
     this.timerSubscription = observableTimer(1000, this.timerDelay).subscribe(() => {
       this.refreshPage();
+      this.changeStatus();
     });
     console.log(`Timer delay time set to ${this.timerDelay / 1000} s.`);
   }
@@ -228,7 +235,7 @@ export class ProjectEditorComponent implements OnDestroy, OnInit {
           this.incomeCompareSet["round"][elementItem.Id] = this.incomeCompareSet["round"][elementItem.Id] + 1;
         } else if (before > after) { // decreased
           this.incomeCompareSet["round"][elementItem.Id] = this.incomeCompareSet["round"][elementItem.Id] + 1;
-        } else { // stable
+        } else { // avarage
           this.incomeCompareSet["round"][elementItem.Id] = 0;
         }
 
@@ -242,16 +249,91 @@ export class ProjectEditorComponent implements OnDestroy, OnInit {
         });
 
         // Sort for allRoundsIncome (only for the main)
-        if (this.selectedElement.Id === mainElement.Id){
+        if (this.selectedElement === mainElement)
           this.selectedElement.getElementItemSet("allRoundsIncome");
-        }
 
       });
     });
 
+  }
+
+  // Change selected item if parent item is decreasing
+  changeStatus(): void {
+
+    if (this.paused) {
+      return;
+    }
+
+    this.project.ElementSet.forEach((element) => {
+      element.ElementItemSet.forEach((elementItem) => {
+        var before = this.incomeCompareSet["before"][elementItem.Id]["income"];
+        var after = this.incomeCompareSet["after"][elementItem.Id]["income"];
+
+        if (before < after) { // increased
+
+          // Increased income item removes array from decreasedItems
+          if (element === this.project.ElementSet[0]) {
+            if (this.decreasedItems.indexOf(elementItem) > -1)
+              this.decreasedItems.splice(this.decreasedItems.indexOf(elementItem), 1);
+            }
+
+          } else if (before > after) { // decrease
+
+          // Item decreases when i try to change item selectedItemElement (only for the main)
+          if (element === this.project.ElementSet[0]) {
+            if (this.decreasedItems.indexOf(elementItem) === -1)
+              this.decreasedItems.push(elementItem);
+          }
+
+        }
+      });
+    });
+
     // When item income decreases tried to change the status
-    //if (this.decreasedItems.length > 0) this.chanceSelectedElementItem();
-  };
+    if (this.decreasedItems.length > 0) this.chanceSelectedElementItem();
+  }
+
+  // Change parent element selected item
+  // If random number is greater then seven.
+  chanceSelectedElementItem(): void {
+
+    const r = Math.floor(Math.random() * this.decreasedItems.length);
+    const random = Math.floor(Math.random() * 10);
+    //this.decreasedItems.forEach((x, i) => console.log(i, x.Name));
+
+    if (random > 7) {
+      var item = this.decreasedItems[r];
+      var elementSet = this.project.ElementSet
+      var elementCellSet = item.ElementCellSet;
+
+      console.log(" -- Parent Element:", item.Name);
+      if (elementSet.length > 0) {
+
+        if (elementCellSet.length > 0) {
+          for(var i = 0; i < elementCellSet.length; i++) {
+            var selectedItem = item.ElementCellSet[i].SelectedElementItem;
+            console.log("Selected Item: ",selectedItem.Name);
+            item.ElementCellSet[i].SelectedElementItem = this.randomItem(selectedItem.Element.ElementItemSet, selectedItem);
+            console.log("new Selected Item: ", item.ElementCellSet[i].SelectedElementItem.Name);
+          }
+          // Item removes array from decreasedItems
+          this.decreasedItems.splice(this.decreasedItems.indexOf(item), 1);
+        }
+      }
+
+      //console.log("Length:", this.decreasedItems.length);
+    }
+  }
+
+  // Generate random item form ElementItem array but exclude current selectedElementItem!
+  randomItem(elementItemSet:ElementItem[], item: ElementItem): ElementItem {
+    var rand = Math.floor(Math.random() * elementItemSet.length);
+    if (elementItemSet[rand] === item) {
+      return this.randomItem(elementItemSet, item);
+    } else {
+      return elementItemSet[rand];
+    }
+  }
 
   getIncomeCompareStatus(elementItem: number): string {
     if (this.project.rounds === 0) this.incomeCompareSetInit();
@@ -431,7 +513,7 @@ export class ProjectEditorComponent implements OnDestroy, OnInit {
   }
 
   // if parent element cells selected item then true
-  isSelectedElementItem(elementItem: ElementItem): Boolean {
+  isSelectedElementItem(elementItem: ElementItem, elementCell?: ElementCell): Boolean {
     var isSelectedElementItem = false;
     const mainElement = this.project.ElementSet[0]; // Main (Parent) Element
 
@@ -449,9 +531,11 @@ export class ProjectEditorComponent implements OnDestroy, OnInit {
         }
       });
     }
-
     // This role is only for parent element, other elements return as isSelectedElementItem
-    return this.selectedElement === mainElement ? true : isSelectedElementItem;
+    var result = this.selectedElement === mainElement ? true : isSelectedElementItem;
+    //TODO: check! - if result is false (item rating buttons are disabled) maybe rating set zero
+    //if (!result && elementCell !== null) { this.updateElementCellDecimalValue(elementCell, 0)
+    return result;
   }
 
   // Remove element item